SQL_Q&A/requirements.txt
streamlit
langchain
sqlalchemy
python-dotenv
mysql-connector-python
langchain-community
langchain_groq